Decon

+++ code/game/objects/items/bodybag.dm
@@ -11,11 +11,14 @@
 	var/obj/item/reagent_containers/syringe/syringe
 	var/cryogenic = FALSE
 	var/robotic = FALSE
+	var/mass_grave = FALSE //CHOMPEdit
 
 /obj/item/bodybag/attack_self(mob/user)
 	. = ..(user)
 	if(.)
 		return TRUE
+	if(mass_grave) //CHOMPedit - TODO, upport this.
+		return FALSE //CHOMPedit - TODO, upport this.
 
 	if(cryogenic)
 		var/obj/structure/closet/body_bag/cryobag/R = new /obj/structure/closet/body_bag/cryobag(user.loc)
@@ -64,8 +67,12 @@
 	desc = "A large folded bag designed for the storage and transportation of cadavers."
 	icon = 'icons/obj/closets/bodybag_large.dmi'
 	w_class = ITEMSIZE_LARGE
+	mass_grave = TRUE
 
 /obj/item/bodybag/large/attack_self(mob/user)
+	. = ..(user)
+	if(.)
+		return TRUE
 	var/obj/structure/closet/body_bag/large/R = new /obj/structure/closet/body_bag/large(user.loc)
 	R.add_fingerprint(user)
 	qdel(src)
